Academic Project - Time Series analysis gold price
Solving for Golden Harvest
Golden Harvest is an upcoming startup that provides customised gold lockets and charms. One of the major problems that it has been facing is associated with the volatility in the price of gold. In order to plan for the financial year better and to look at the possibility of a fixed price offering, they have hired you as a consultant.

You have been provided with the data from the year 1992 to 2019. Here we will be using the data from 1992 to 2017 to build the forecasting model and we will test this model over the data from 2018-2019.